+++ public/client.mjs
@@ -1,6 +1,5 @@
 // console.log('Loading');
-const log = console.log.bind(console);
-
+import { log } from "./lib/debug.mjs";
 import "./lib/gui.mjs";
 import {Message} from "./lib/message.mjs";